ml/requirements.txt
numpy>=1.26
pandas>=2.2
Pillow>=10.4
psycopg2-binary>=2.9
PyYAML>=6.0
requests>=2.32
scikit-learn>=1.5
torch>=2.4
torchvision>=0.19
onnx>=1.16
onnxruntime>=1.18
